requirements.txt
PyQt5==5.15.9
mysql-connector-python==8.1.0
bcrypt==4.0.1
cryptography==41.0.3
